authentication

+++ app/Http/Controllers/TasksController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Models\Task;    // 追加
+use Illuminate\Support\Facades\Auth;
 
 class TasksController extends Controller
 {
@@ -18,12 +19,17 @@ class TasksController extends Controller
     public function index()
     {
         // メッセージ一覧を取得
-        $tasks = Task::all();         // 追加
-
+        $tasks = Task::all();
+        
+        if (Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = Auth::user();
+        }
+        
         // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [     // 追加
-            'tasks' => $tasks,        // 追加
-        ]);                                 // 追加
+        return view('tasks.index', [
+            'tasks' => $tasks,
+        ]);
     }
 
     /**
@@ -53,6 +59,7 @@ class TasksController extends Controller
     // postでtasks/にアクセスされた場合の「新規登録処理」
     public function store(Request $request)
     {
+        // dd(\Auth::id());
         // バリデーション
         $request->validate([
             'status' => 'required|max:10',
@@ -63,10 +70,11 @@ class TasksController extends Controller
         $task = new Task;
         $task->status = $request->status;
         $task->content = $request->content;
+        $task->user_id = Auth::id();
         $task->save();
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // タスク一覧ページへリダイレクトさせる
+        return redirect('/tasks');
     }
 
     /**
@@ -83,9 +91,14 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // メッセージ詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        
+        // トップページへリダイレクトさせる
+        return redirect('/');
     }
 
     /**
@@ -102,9 +115,14 @@ class TasksController extends Controller
         $task = Task::findOrFail($id);
 
         // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        
+        // トップページへリダイレクトさせる
+        return redirect('/');
     }
 
     /**
@@ -127,9 +145,13 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
         // メッセージを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        if (\Auth::id() === $task->user_id) {
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+            // タスク一覧ページへリダイレクトさせる
+            return redirect('/tasks');
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -148,8 +170,12 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task =Task::findOrFail($id);
         // メッセージを削除
-        $task->delete();
-
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            // タスク一覧ページへリダイレクトさせる
+            return redirect('/tasks');
+        }
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
